# swagger.yaml
openapi: 3.0.0
info:
  title: Online JobFair API
  description: A secure Node.js Express backend API for managing job fair operations, including companies, interviews, and user reviews.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
  - url: https://se-project-be-68-2-anumafia.vercel.app/api/v1
    description: Production Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Company:
      type: object
      required: [user, name, tel]
      properties:
        id: { type: string }
        user: { type: string, description: "User ID of the company owner" }
        name: { type: string }
        tel: { type: string }
        public: { type: boolean, default: false }
        ratingAverage: { type: number, minimum: 0, maximum: 5 }
        reviewCount: { type: integer }
        address: { type: string }
        description: { type: string }
        website: { type: string, format: url }
...
    Review:
      type: object
      required: [user, company, rating, reviewText]
      properties:
        id: { type: string }
        user: { type: string, description: "User ID" }
        company: { type: string, description: "Company ID" }
        rating: { type: integer, minimum: 0, maximum: 5 }
        reviewText: { type: string }
        createdAt: { type: string, format: date-time }

    # Required for paths to resolve correctly
...
security:
  - bearerAuth: []
paths:
...
  /companies:
    get:
      tags: [Companies]
      summary: Get all companies
      security: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: select
          schema: { type: string }
      responses:
        '200': { description: Array of companies }
    post:
      tags: [Companies]
      summary: Create a company (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  properties:
                    name: { type: string }
                    tel: { type: string }
                    email: { type: string, format: email }
                    password: { type: string, format: password }
                    role: { type: string, enum: [company] }
                - $ref: '#/components/schemas/Company'
      responses:
        '201': { description: Company created }
...
  /reviews:
    get:
      tags: [Reviews]
      summary: Get all reviews (Admin)
      responses:
        '200': { description: Array of all reviews }
    post:
      tags: [Reviews]
      summary: Create a review for a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201': { description: Review created }
...
  /users:
    get:
      tags: [Users]
      summary: Get all users (Admin only)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200': { description: Paginated array of users }
...